56 task 13 04 smart home heater shh

+++ SmartHomeSystem/backend/src/main/java/observer/Observer.java
@@ -1,5 +1,7 @@
 package observer;
 
+
+// Define the Observer interface
 public interface Observer {
-    void update(int temperature);
+    void update(Double temperature);
 }
